+++ sujog-migration/src/main/java/org/egov/migration/config/BatchConfiguration.java
@@ -5,10 +5,10 @@ import java.io.IOException;
 import org.apache.poi.EncryptedDocumentException;
 import org.egov.migration.business.model.PropertyDetailDTO;
 import org.egov.migration.common.model.RecordStatistic;
-import org.egov.migration.model.Property;
 import org.egov.migration.processor.PropertyItemWriter;
 import org.egov.migration.processor.PropertyReader;
 import org.egov.migration.processor.PropertyTransformProcessor;
+import org.egov.migration.reader.model.Property;
 import org.springframework.batch.core.Step;
 import org.springframework.batch.core.configuration.annotation.EnableBatchProcessing;
 import org.springframework.batch.core.configuration.annotation.JobBuilderFactory;
