+++ sujog-migration/src/main/java/org/egov/migration/mapper/DemandDetailRowMapper.java
@@ -3,7 +3,7 @@ package org.egov.migration.mapper;
 import java.util.Map;
 
 import org.apache.poi.ss.usermodel.Row;
-import org.egov.migration.model.DemandDetail;
+import org.egov.migration.reader.model.DemandDetail;
 import org.egov.migration.util.MigrationConst;
 import org.egov.migration.util.MigrationUtility;
 import org.springframework.stereotype.Component;
