+++ sujog-migration/src/main/java/org/egov/migration/mapper/DemandRowMapper.java
@@ -3,7 +3,7 @@ package org.egov.migration.mapper;
 import java.util.Map;
 
 import org.apache.poi.ss.usermodel.Row;
-import org.egov.migration.model.Demand;
+import org.egov.migration.reader.model.Demand;
 import org.egov.migration.util.MigrationConst;
 import org.egov.migration.util.MigrationUtility;
 import org.springframework.stereotype.Component;
