+++ sujog-migration/src/main/java/org/egov/migration/processor/PropertyReader.java
@@ -22,13 +22,13 @@ import org.egov.migration.mapper.DemandRowMapper;
 import org.egov.migration.mapper.OwnerRowMapper;
 import org.egov.migration.mapper.PropertyRowMapper;
 import org.egov.migration.mapper.PropertyUnitRowMapper;
-import org.egov.migration.model.Address;
-import org.egov.migration.model.Assessment;
-import org.egov.migration.model.Demand;
-import org.egov.migration.model.DemandDetail;
-import org.egov.migration.model.Owner;
-import org.egov.migration.model.Property;
-import org.egov.migration.model.PropertyUnit;
+import org.egov.migration.reader.model.Address;
+import org.egov.migration.reader.model.Assessment;
+import org.egov.migration.reader.model.Demand;
+import org.egov.migration.reader.model.DemandDetail;
+import org.egov.migration.reader.model.Owner;
+import org.egov.migration.reader.model.Property;
+import org.egov.migration.reader.model.PropertyUnit;
 import org.egov.migration.util.MigrationConst;
 import org.egov.migration.util.MigrationUtility;
 import org.springframework.batch.item.ItemReader;
