+++ sujog-migration/src/main/java/org/egov/migration/processor/PropertyTransformProcessor.java
@@ -17,10 +17,10 @@ import org.egov.migration.business.model.PropertyDTO;
 import org.egov.migration.business.model.PropertyDetailDTO;
 import org.egov.migration.business.model.UnitDTO;
 import org.egov.migration.config.SystemProperties;
-import org.egov.migration.model.Address;
-import org.egov.migration.model.Assessment;
-import org.egov.migration.model.DemandDetail;
-import org.egov.migration.model.Property;
+import org.egov.migration.reader.model.Address;
+import org.egov.migration.reader.model.Assessment;
+import org.egov.migration.reader.model.DemandDetail;
+import org.egov.migration.reader.model.Property;
 import org.egov.migration.service.ValidationService;
 import org.egov.migration.util.MigrationConst;
 import org.egov.migration.util.MigrationUtility;
