+++ sujog-migration/src/main/java/org/egov/migration/reader/model/Address.java
@@ -1,4 +1,4 @@
-package org.egov.migration.model;
+package org.egov.migration.reader.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
