+++ sujog-migration/src/main/java/org/egov/migration/reader/model/Owner.java
@@ -1,4 +1,4 @@
-package org.egov.migration.model;
+package org.egov.migration.reader.model;
 
 import javax.validation.constraints.Email;
 import javax.validation.constraints.NotEmpty;
