+++ sujog-migration/src/main/java/org/egov/migration/reader/model/Property.java
@@ -1,4 +1,4 @@
-package org.egov.migration.model;
+package org.egov.migration.reader.model;
 
 import java.util.List;
