+++ sujog-migration/src/main/java/org/egov/migration/reader/model/PropertyUnit.java
@@ -1,4 +1,4 @@
-package org.egov.migration.model;
+package org.egov.migration.reader.model;
 
 import java.math.BigDecimal;
